requirements.txt
opencv-python==4.5.5.64
numpy==1.22.3
PyQt5==5.15.6
pywin32==304
comtypes==1.1.14
Pillow==9.3.0
mediapipe==0.9.0.1
absl-py>=0.12.0
